Add compiler class
After reviewing the literature on lexers, parsers, and compilers, the Parser class needs to be renamed to Lexer. A Lexer tokenizes the input. A Parser converts tokens to syntax, and a compiler embodies a lexer and parser. So, the Compiler class is the root and should have a lexer object and a parser object. It should also have an encoder object/class that converts the parser output (syntax) to a string or file. The parser would check that all sections begin and end properly, but I am not sure what other functions it would perform. There is not much grammer/logic to the mustache specification, by design. So, it might be possible to just include the parser as part of the encoder.

A Writer class might be needed too.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@55 e755202d-4ee8-5f4b-928a-6a4ca417d9fe