Change Reader and Writer libraries to low-level
The user of this toolkit only needs to be aware of the Reader and Writer classes and libraries if he or she is trying to implement his or her own. Most of the time, the usage of these classes should be invisible to the user. The Data library is a different story.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@82 e755202d-4ee8-5f4b-928a-6a4ca417d9fe