Add tokenizing comments
Previously comments were not added to the list of tokens during texting. Instead, a comment token should be added to the list and the parsing should add it to the abstract syntax tree and the compiler decide what to do with the comments. This way, future compiles can do something with the comments if needed.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@98 e755202d-4ee8-5f4b-928a-6a4ca417d9fe