Remove unnecessary private member and accessor VIs
The parent private member for nodes is not needed. In fact, this causes multiple copies of the same root or parent object to exist in the tree. We never reverse up the tree from a child to the parent. We only walk down the tree, so the parent private member and accessors are unnecessary. If this feature is ever needed, Data Value References (DVRs) will be needed to avoid copies of the parent object.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@108 e755202d-4ee8-5f4b-928a-6a4ca417d9fe