Change implementation to not use "Content"/"Output" VI method within node class
The "Content" method VI for nodes was a little weird. The nodes were responsible for rendering themselves, but in reality a "Render" should act on the nodes. The nodes should not know how to render themselves, but be a container of the information necessary for a render to extract the output. This should make creating the "source", or original content from the nodes easier and the intent more clear.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@112 e755202d-4ee8-5f4b-928a-6a4ca417d9fe