Add Read Type VI for data classes
The Read Type VI makes implementation more robust to changes in project library names and avoids using strings to identify object class with the Get LV Class VI. In theory, the project library name should never change, so using the fully qualified string name of a class should be sufficient, but typos in the names and possible refactoring the names could leave to bugs and failed tests. An enum and type definition is used to avoid using strings for type identification.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@164 e755202d-4ee8-5f4b-928a-6a4ca417d9fe