Add workaround for the Get LV Class Name function not existing for LabVIEW 2009
Basically, using the Get LV Class Name function was kind of brittle anyways. Using string is problematic with typos and name changes. So, a typedef enum is used with a getter method for the type. A string that returns the class name could have been implemented, but that would not eliminate problems with typos and name changes.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@210 e755202d-4ee8-5f4b-928a-6a4ca417d9fe