Add Pair data class to workaround compile error with LabVIEW 2009
Apparently, LabVIEW 2009 does not like arrays of maps to be used to create a map. I think this is a recursive VI and causes problems during compiling. This recursion is allowed in a newer versions, but a workaround is created with a new Data class, which is more OOP anyways.

git-svn-id: svn://127.0.0.1/Mustache-LV/trunk@212 e755202d-4ee8-5f4b-928a-6a4ca417d9fe