Add descriptions to type definitions and descriptions to the controls within the type definitions
The NI reviewer came back saying the VI descriptions for the API was incomplete. I double checked everything in the VIs before submitting the package for review. I am not sure which ones the reviewer is referring to. I have requested more specific information. In the mean time, all I could find were missing descriptions for the type definitions.