Add more tests to investigate possible infinite loop
See #4. I had some tests, but I am unable at the moment to re-create the infinite loop observed in the other project related to partials and recursion. I may need to tweak existing tests and/or work a little bit more on re-creating the conditions. I might have to debug some in the other project.