Add render process hooks
Pre and post hooks for the rendering "process" are added. This is known as the "Process" hook under the new `Render` namespace library. These are executed before and after rendering. This can be useful for setting up, initializing, starting, tearing down, shutting down, stopping, etc. hooks and other VIs when the rendering is started or completed, respectively.