Change `Context` class to a public, but low-level class
The `Context` class is made public and is the container for hooks instead of a cluster. This feels like an improvement over a big cluster of hook objects. This also feels more OOP. However, this is just the initial implementation that only focuses on the Pre/Post compile process hook. Read/Write accessor methods will have to be created for all of the hooks in the Context class. Also, there is still a question on the Context API. Should there be "Create/Destroy" VIs? How is a default context handled?