Change all stages of compiling to classes
The organization is completely changed so that there is a Lexer, Parser, and Renderer classes that are components of the Compiler class. The Context is moved to the Renderer, where it is needed. Everything is re-organized to be within one of these four classes for compiling a template. There are still syntax errors and on-disk organization issues, but this is the bulk of the experimental implementation, which allows a user to swap out a lexer, a parser, or a renderer for a compiler. This makes the library much more flexible and extensible. Hooks are moved, but not complete yet, to specific classes.