Change data access VIs to require/optional inputs for Initialize VIs
A Lexer and Renderer are required to have a Reader and Writer, respectively. Before, data access VIs were used, `Write Reader.vi` and `Write Writer.vi`, but _must_ be used "attaching" the lexer and renderer to the compiler, respectively. It is possible with this implementation for a developer to _not_ attach a reader and writer, and an error will result. Rather than assume the developer has read the documentation and called these VIs, the Reader and Writer are moved to required inputs for the Initialize.vi of each respective class. This enforces the requirement to create a reader and writer for each of the objects rather than relying on the documentation. The data access members are removed.